♻️메인 페이지 SSR 방식으로 변경
